Feat: Add Dancepace Scriptable Object
