+++ Assets/Resources/SO.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c71ccf06f7e59e1449cd26366a9a0276
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
