+++ Assets/Resources/SO/Dancepace.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5ed8401ba418e946a2f0ad3d2df4e1f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
